KUBE-17, KUBE-26, KUBE-27: search connectivity tool, bootstrap test mixins, background availability tester
- tests/common/search/connectivity.py — SearchConnectivityTool ($search
  one-shot + paging), failure-class taxonomy (cursor_lost /
  transient_network / other), pod/STS disruption helpers
  (delete_pods, wait_for_all_pods_replaced, wait_for_pods_by_label_replaced,
  wait_for_mongot_statefulset_drained, patch_mongot_readiness_probe_to_false).
- tests/common/search/bootstrap_test_mixins.py — layered reusable mixins
  for the deploy → search → data → smoke bootstrap (RS, sharded, MC).
- tests/common/search/background_availability_tester.py — daemon-thread
  outage-window verdict accumulator.
- tests/search/search_connectivity_tool*.py + search_failure_modes.py —
  three new e2e markers exercising mongot outage, pod restart, envoy
  restart, graceful shutdown, sharded fanout disruption, and background-
  tester pod-restart scenarios.
- kubetester/__init__.py — promoted list_matching_pods / pod_is_ready
  to public; added wait_for_pods_ready / wait_for_no_pods_ready.
- .evergreen.yml / .evergreen-tasks.yml — register the three markers
  inside e2e_mdb_kind_search_task_group.

+++ docker/mongodb-kubernetes-tests/kubetester/mongotester.py
@@ -600,7 +600,7 @@ class BackgroundHealthChecker(threading.Thread):
     def __init__(
         self,
         health_function,
-        wait_sec: int = 3,
+        wait_sec: float = 3,
         allowed_sequential_failures: int = 3,
         health_function_params=None,
     ):
@@ -686,7 +686,7 @@ class MongoDBBackgroundTester(BackgroundHealthChecker):
     def __init__(
         self,
         mongo_tester: MongoTester,
-        wait_sec: int = 3,
+        wait_sec: float = 3,
         allowed_sequential_failures: int = 1,
         health_function_params=None,
     ):
